+++ src/components/layout/Testimonials.jsx
@@ -1,15 +1,10 @@
 import Carousel from 'better-react-carousel';
-import clsx from 'clsx';
 import { Tweet } from 'react-twitter-widgets';
 
 const Testimonials = ({ title = 'Testimonials', tweetId }) => {
   return (
     <section id='testimonials'>
-      <div
-        className={clsx(
-          'layout flex flex-col items-center justify-center gap-6 py-20'
-        )}
-      >
+      <div className='layout flex flex-col items-center justify-center gap-6 py-20'>
         <h2 className='h1'>{title}</h2>
         <hr className='styled-hr' />
         <Carousel
